easyCurl re-structure
- Draw the curl function off from easyCurl to easyCurlKernel.
- Move easyCurlKernel and easyCurlConfig to subfolder kernel
- Move clsRest, clsSum and easyCurlTest to subfolder wrap
