+++ src/kernel/easyCurlConfig.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 
 # REST METHOD SUPPORTED BY EASYCURL, OF COURSE YOU CAN ADD MORE
-optCurlMethods = {'COMMON': [('pycurl.TIMEOUT', 'int', '10'),
+optCurlMethods = {'COMMON': [('pycurl.TIMEOUT', 'int', '100'),
                              ('pycurl.VERBOSE', 'bool', 'False')],
                   'GET'   : [('pycurl.HTTPGET', 'int', '1')],
                   'POST'  : [('pycurl.POST', 'int', '1'),
